requirements.txt
mysql-connector-python
scrapy
dateparser
python-dotenv
beautifulsoup4
lxml
emoji